requirements.txt
# Core ML & Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# PDF Generation
reportlab>=4.0.0
weasyprint>=60.0

# Utilities
joblib>=1.3.0
python-dotenv>=1.0.0

# Validation
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Production
waitress>=2.1.0
